requirements.txt
twitchio>=2.6.0
pyyaml~=6.0.1